requirements.txt
accelerate===1.12.0
datasets===4.4.1
dramatiq[redis]===1.17.0
python-dotenv===1.0.1
fastapi===0.123.8
numpy<2
peft===0.18.0
protobuf===5.29.2
pydantic===2.12.5
redis===5.0.1
sentencepiece===0.2.0
torch===2.2.2
transformers===4.57.3
trl===0.25.1
uvicorn[standard]===0.34.0
